+++ helpers/firefox_cookie.py
